requirements.txt
tensorflow-cpu==2.10.0
flask>=2.0
pillow>=9.0
numpy>=1.0
gunicorn>=20.0